Use our own display brightness icon
We use the "virransäästöliitännäinen" xfce4 panel plugin to adjust
display brightness. We used to fork the whole xfce4 just to remove
non-display-brighness stuff from the plugin UI but now we use the
vanilla plugin with some extra UI elements to reduce hacks.

It uses a bunch of different icons depending on battery state so we
override them all with our own icon.